Move methods to util class to reduce clutter
